Adding diagnostic workflow + fixes
