+++ earth2studio/models/dx/climatenet.py
@@ -81,7 +81,7 @@ class ClimateNet(torch.nn.Module, AutoModelMixin):
 
     input_coords = OrderedDict(
         {
-            "batch": np.empty(1),
+            "batch": np.empty(0),
             "variable": np.array(VARIABLES),
             "lat": np.linspace(90, -90, 721, endpoint=True),
             "lon": np.linspace(90, -90, 1440, endpoint=False),
@@ -90,7 +90,7 @@ class ClimateNet(torch.nn.Module, AutoModelMixin):
 
     output_coords = OrderedDict(
         {
-            "batch": np.empty(1),
+            "batch": np.empty(0),
             "variable": np.array(OUT_VARIABLES),
             "lat": np.linspace(90, -90, 721, endpoint=True),
             "lon": np.linspace(90, -90, 1440, endpoint=False),
