+++ earth2studio/models/dx/corrdiff.py
@@ -137,7 +137,7 @@ class CorrDiffTaiwan(torch.nn.Module, AutoModelMixin):
     def input_coords(self) -> CoordSystem:
         return OrderedDict(
             {
-                "batch": np.empty(1),
+                "batch": np.empty(0),
                 "variable": np.array(VARIABLES),
                 "lat": np.linspace(19.25, 28, 36, endpoint=True),
                 "lon": np.linspace(116, 126, 40, endpoint=False),
@@ -148,7 +148,7 @@ class CorrDiffTaiwan(torch.nn.Module, AutoModelMixin):
     def output_coords(self) -> CoordSystem:
         return OrderedDict(
             {
-                "batch": np.empty(1),
+                "batch": np.empty(0),
                 "sample": np.arange(self.number_of_samples),
                 "variable": np.array(OUT_VARIABLES),
                 "ilat": np.arange(448),
