+++ earth2studio/models/px/pangu.py
@@ -132,7 +132,7 @@ class PanguBase(torch.nn.Module, AutoModelMixin, PrognosticMixin):
 
         self._input_coords = OrderedDict(
             {
-                "batch": np.empty(1),
+                "batch": np.empty(0),
                 "lead_time": np.array([np.timedelta64(0, "h")]),
                 "variable": np.array(VARIABLES),
                 "lat": np.linspace(90, -90, 721, endpoint=True),
@@ -142,7 +142,7 @@ class PanguBase(torch.nn.Module, AutoModelMixin, PrognosticMixin):
 
         self._output_coords = OrderedDict(
             {
-                "batch": np.empty(1),
+                "batch": np.empty(0),
                 "lead_time": np.array([np.timedelta64(6, "h")]),
                 "variable": np.array(VARIABLES),
                 "lat": np.linspace(90, -90, 721, endpoint=True),
