+++ earth2studio/utils/coords.py
@@ -272,21 +272,21 @@ def map_coords(
     return x, mapped_coords
 
 
-def extract_coords(
+def split_coords(
     x: torch.Tensor, coords: CoordSystem, dim: str = "variable"
 ) -> tuple[list[torch.Tensor], CoordSystem, np.ndarray]:
     """
-    A utility function to extract a dimension from a (x,coords) pair and convert it into a list of tensors, a CoordSystem, and
-    the dimension that extract from coords.
+    A utility function to split a dimension from a (x,coords) pair and convert it into
+    a list of tensors, a CoordSystem, and the dimension that extract from coords.
 
     Parameters
     ----------
     x : torch.Tensor
         Input tensor
     coords : CoordSystem
-        Coordinates referring to the dimensions of x.
+        Coordinates referring to the dimensions of x
     dim : str
-        Name of the dimension in coords to extract by.
+        Name of the dimension in coords to split along
 
     Returns
     -------
