+++ examples/extend/01_custom_prognostic.py
@@ -68,7 +68,7 @@ class CustomPrognostic(torch.nn.Module):
 
     input_coords = OrderedDict(
         {
-            "batch": np.empty(1),
+            "batch": np.empty(0),
             "lead_time": np.array([np.timedelta64(0, "h")]),
             "variable": np.array(["u10m", "v10m"]),
             "lat": np.linspace(90, -90, 721),
@@ -78,7 +78,7 @@ class CustomPrognostic(torch.nn.Module):
 
     output_coords = OrderedDict(
         {
-            "batch": np.empty(1),
+            "batch": np.empty(0),
             "lead_time": np.array([np.timedelta64(1, "h")]),
             "variable": np.array(["u10m", "v10m"]),
             "lat": np.linspace(90, -90, 721),
@@ -193,11 +193,6 @@ class CustomPrognostic(torch.nn.Module):
 # ------
 # With the custom prognostic defined, it's now easily usable in a standard workflow. In
 # this example, we will use the build in workflow :py:meth:`earth2studio.run.deterministic`.
-
-# %%
-# .. literalinclude:: ../../earth2studio/run.py
-#    :language: python
-#    :lines: 35-42
 
 # %%
 # Let's instantiate the components needed.
