+++ test/io/test_kv.py
@@ -23,7 +23,7 @@ import torch
 import xarray as xr
 
 from earth2studio.io import KVBackend
-from earth2studio.utils.coords import extract_coords
+from earth2studio.utils.coords import split_coords
 
 
 @pytest.mark.parametrize(
@@ -160,7 +160,7 @@ def test_kv_variable(
     )
     partial_data = torch.randn((1, 1, 180, 180), device=device)
 
-    z.write(*extract_coords(partial_data, partial_coords, "variable"))
+    z.write(*split_coords(partial_data, partial_coords, "variable"))
     assert torch.allclose(z[variable[0]][0, :, :180], partial_data)
 
     # test to xarray
