+++ test/io/test_netcdf4.py
@@ -25,7 +25,7 @@ import pytest
 import torch
 
 from earth2studio.io import NetCDF4Backend
-from earth2studio.utils.coords import extract_coords
+from earth2studio.utils.coords import split_coords
 
 
 @pytest.mark.parametrize(
@@ -225,7 +225,7 @@ def test_netcdf4_variable(
         }
     )
     partial_data = torch.randn((1, 1, 180, 180), device=device)
-    nc.write(*extract_coords(partial_data, partial_coords, "variable"))
+    nc.write(*split_coords(partial_data, partial_coords, "variable"))
     assert np.allclose(nc[variable[0]][0, :, :180], partial_data.to("cpu").numpy())
     nc.close()
 
@@ -255,7 +255,7 @@ def test_netcdf4_variable(
             }
         )
         partial_data = torch.randn((1, 1, 180, 180), device=device)
-        nc.write(*extract_coords(partial_data, partial_coords, "variable"))
+        nc.write(*split_coords(partial_data, partial_coords, "variable"))
         assert np.allclose(nc[variable[0]][0, :, :180], partial_data.to("cpu").numpy())
         nc.close()
