+++ test/io/test_zarr.py
@@ -25,7 +25,7 @@ import torch
 import zarr
 
 from earth2studio.io import ZarrBackend
-from earth2studio.utils.coords import extract_coords
+from earth2studio.utils.coords import split_coords
 
 
 @pytest.mark.parametrize(
@@ -238,7 +238,7 @@ def test_zarr_variable(
         }
     )
     partial_data = torch.randn((1, 1, 180, 180), device=device)
-    z.write(*extract_coords(partial_data, partial_coords, "variable"))
+    z.write(*split_coords(partial_data, partial_coords, "variable"))
     assert np.allclose(z[variable[0]][0, :, :180], partial_data.to("cpu").numpy())
 
     # Test Directory Store
@@ -271,7 +271,7 @@ def test_zarr_variable(
             }
         )
         partial_data = torch.randn((1, 1, 180, 180), device=device)
-        z.write(*extract_coords(partial_data, partial_coords, "variable"))
+        z.write(*split_coords(partial_data, partial_coords, "variable"))
         assert np.allclose(z[variable[0]][0, :, :180], partial_data.to("cpu").numpy())
